+++ local-addons/supco/security/ir.model.access.csv
