+++ local-addons/supco/views/supco_department_views.xml
